requirements.txt
# RAG AI Chatbot — Academic City
# CS4241 Introduction to Artificial Intelligence 2026
# Student: [Jacqueline Naa Ayima Mensah] | Index: [10022200119]

# ── Core Data ──────────────────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0

# ── PDF Parsing ────────────────────────────────────────────────────────────────
PyMuPDF>=1.23.0          # Fast, accurate PDF text extraction (used in Part A)

# ── NLP / Tokenization ────────────────────────────────────────────────────────
nltk>=3.8.1              # Sentence tokenization for chunking strategies

# ── HTTP / Downloads ──────────────────────────────────────────────────────────
requests>=2.31.0
tqdm>=4.65.0             # Progress bars for downloads

# ── Embeddings (Part B) ───────────────────────────────────────────────────────
sentence-transformers>=2.6.0   # Local sentence embeddings — no API key needed
faiss-cpu>=1.7.4               # Vector store for similarity search

# ── Visualisation (Part A analysis) ──────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0

# ── LLM (Part C / D) ─────────────────────────────────────────────────────────
openai>=1.10.0           # or use a free alternative like groq/together

# ── UI (Final Deliverable) ────────────────────────────────────────────────────
streamlit>=1.33.0
